Made sure that bounding boxes are reset when mode or circuits are changed.
Previously there was a bug that when bboxes are selected and circuits are changed the bboxes are still saved in the background so it was a problem that when new bboxes in the new circuit are selected it provided an error that only 2 elements can be selected. So you had to manually reset the bboxes. Now this process is automated, so that whenever changes in the menu occur it automatically clears the remaining selected bboxes.