Added null-handler for jsonFilePath_VC in the display steps function, so that from now on the emission of this filepath in the header doesnt cause any more issues
This means, that e.g. for the handling of the previous and next buttons, the respective display_steps headers dont need corresponding json files for the VC calculations. That would mean, that a "crash" would only be encountered once the "next" button is triggered, because only then would the program truly need the json_vc files. Whether it is necessary to further eliminate the necessity for the json_vc files even during the vc calculations (so when the next button is pressed) seems debatable, but could be achieved...